requirements.txt
Django==5.0.6
djangorestframework==3.14.0
psycopg2-binary
